+++ src/meltano/core/plugin_remove_service.py
@@ -18,7 +18,12 @@ class PluginRemoveService:
     """Handle plugin installation removal operations."""
 
     def __init__(self, project: Project, plugins_service: ProjectPluginsService = None):
-        """Construct a PluginRemoveService instance."""
+        """Construct a PluginRemoveService instance.
+
+        Args:
+            project: The Meltano project.
+            plugins_service: The project plugins service.
+        """
         self.project = project
         self.plugins_service = plugins_service or ProjectPluginsService(project)
 
@@ -34,6 +39,16 @@ class PluginRemoveService:
         Returns a tuple containing:
         1. The total number of removed plugins
         2. The total number of plugins attempted
+
+        Args:
+            plugins: The plugins to remove.
+            plugin_status_cb: A callback to call for each plugin.
+            removal_manager_status_cb: A callback to call for each removal manager.
+
+        Returns:
+            A tuple containing:
+            1. The total number of removed plugins
+            2. The total number of plugins attempted
         """
         num_plugins: int = len(plugins)
         removed_plugins: int = num_plugins
@@ -54,10 +69,16 @@ class PluginRemoveService:
         return removed_plugins, num_plugins
 
     def remove_plugin(self, plugin: ProjectPlugin) -> Tuple[PluginLocationRemoveStatus]:
-        """
-        Remove a plugin from `meltano.yml`, its installation in `.meltano`, and its settings in the Meltano system database.
+        """Remove a plugin.
 
-        Returns a tuple containing a remove manager for each location.
+        Removes from `meltano.yml`, its installation in `.meltano`, and its settings in
+        the Meltano system database.
+
+        Args:
+            plugin: The plugin to remove.
+
+        Returns:
+            A tuple containing a remove manager for each location.
         """
         remove_managers = (
             DbRemoveManager(plugin, self.project),