+++ tests/fixtures/core.py
@@ -87,7 +87,7 @@ def discovery():  # noqa: WPS213
                             "description": "Run tests",
                         },
                         "test_extra": {
-                            "args": None,
+                            "args": "test_extra",
                             "description": "Run extra tests",
                             "executable": "test-extra",
                         },
@@ -737,5 +737,8 @@ def state_service(job_history_session):
 @pytest.fixture
 def project_with_environment(project: Project) -> Project:
     project.activate_environment("dev")
+    project.active_environment.env[
+        "ENVIRONMENT_ENV_VAR"
+    ] = "${MELTANO_PROJECT_ROOT}/file.txt"
     yield project
     project.active_environment = None